- moved TransactionGuardianIcon and TransactionSovereignBridgeIcon into components/TransactionIcon - renamed TransactionIcon into TransactionStatusIcon - added common component for transaction icons ( TransactionIcon )
