+++ src/components/TransactionIcon/TransactionGuardianIcon/TransactionGuardianIcon.tsx
