+++ src/components/TransactionIcon/TransactionGuardianIcon/index.ts
