+++ src/components/TransactionIcon/TransactionSovereignBridgeIcon/TransactionSovereignBridgeIcon.tsx
