+++ src/components/TransactionIcon/TransactionSovereignBridgeIcon/index.ts
