+++ src/components/TransactionIcon/index.ts
@@ -1,1 +1,4 @@
 export * from './TransactionIcon';
+export * from './TransactionGuardianIcon';
+export * from './TransactionStatusIcon';
+export * from './TransactionSovereignBridgeIcon';
