+++ src/components/index.ts
@@ -68,8 +68,6 @@ export * from './TimeAgo';
 export * from './TokenBlock';
 export * from './Tooltips';
 export * from './TransactionAction';
-export * from './TransactionGuardianIcon';
-export * from './TransactionSovereignBridgeIcon';
 export * from './TransactionIcon';
 export * from './TransactionsTable';
 export * from './TransactionStatus';
